+++ Sources/NerdzAppUpdates/VersionProviders/AppStore/AppStoreErrorResponseModel.swift
@@ -1,6 +1,6 @@
 //
 //  AppStoreErrorResponseModel.swift
-//  
+//
 //
 //  Created by Roman Kovalchuk on 13.07.2021.
 //
@@ -10,8 +10,8 @@ import NerdzNetworking
 
 struct AppStoreErrorResponseModel: Decodable, ServerError {
     let errorMessage: String?
-    
+
     var message: String {
-        return errorMessage ?? "Version check unknown error"
+        errorMessage ?? "Version check unknown error"
     }
 }