+++ Tests/NerdzAppUpdatesTests/TestHelpers/TestData.swift
@@ -1,0 +1,11 @@
+@testable import NerdzAppUpdates
+
+enum TestData {
+    static func success(_ type: AppUpdateType, _ version: String?) -> Result<VersionProviderResult, VersionVerifierError> {
+        .success((type, version))
+    }
+
+    static func failure(_ error: VersionVerifierError) -> Result<VersionProviderResult, VersionVerifierError> {
+        .failure(error)
+    }
+}